requirements.txt
apache-beam==2.27.0
atomicwrites==1.4.0
attrs==20.3.0
avro-python3==1.9.2.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudstorage==0.11.0
colorama==0.4.4
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
dropbox==11.2.0
fastavro==1.3.1
Flask==1.1.2
future==0.18.2
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-bigquery==2.8.0
google-cloud-core==1.6.0
google-cloud-pubsub==2.3.0
google-cloud-storage==1.36.0
google-cloud-vision==2.2.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
GoogleAppEngineCloudStorageClient==1.9.22.1
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.17
MarkupSafe==1.1.1
mock==2.0.0
mocker==1.1.1
mypy-extensions==0.4.3
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
pbr==5.5.1
pluggy==0.13.1
ply==3.11
proto-plus==1.13.0
protobuf==3.14.0
py==1.10.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
pytest-mock==3.5.1
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
stone==3.2.1
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.0
